--mg
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@8581 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/GC/ioa.c
@@ -488,7 +488,7 @@ Program terminated.\n", (int)(4*ReqObjectSize));
   INFO_AOAUSE({
     fprintf(output, "(AOAUSE:%d:%d/%d)\n", 
 	    (int)NumIOAGc, (int)AOAFreeListTotalFree(),
-	    totalAOASize);
+	    (int)totalAOASize);
   });
 } /* End IOAGc */
